cld: Leave rx reorder array indices as int
Leave rx reorder array indices as int rather than u_int16_t
Change re-order indexes to unsigned.

In recent changes to htt_dxe_rx.c, ol_htt_rx_api.h, ol_rx.c, and
ol_rx_reorder.c, the rx reorder indices that are derived from
sequence number are changed from int/unsigned to u_int16_t.
It is reasonable to make these variables consistently unsigned
rather than int in some locations and unsigned in other locations.
However, they should be left as unsigned rather than changed to
u_int16_t, since this per-MPDU code is cycles-critical, and for
a given instruction set and compiler, the generated loops over
rx array elements may be more efficient using unsigned rather
than u_int16_t.

Also, any changes to ol_htt_rx_api.h need to be applied in the
htt/htt_rx.c file as well as in the htt_dxe/htt_dxe_rx.c file.

CL Id: 2034545

+++ cld/CORE/TXRX/ol_rx_reorder.c
@@ -214,7 +214,7 @@ ol_rx_reorder_release(
     adf_nbuf_t tail_msdu;
 
     OL_RX_REORDER_IDX_START_SELF_SELECT(peer, tid, &idx_start);
-    peer->tids_next_rel_idx[tid] = idx_end; /* may get reset below */
+    peer->tids_next_rel_idx[tid] = (u_int16_t)idx_end; /* may get reset below */
 
     win_sz = peer->tids_rx_reorder[tid].win_sz;
     win_sz_mask = peer->tids_rx_reorder[tid].win_sz_mask;
@@ -309,7 +309,7 @@ ol_rx_reorder_flush(
         peer->tids_next_rel_idx[tid] = OL_RX_REORDER_IDX_INIT(
             0/*n/a*/, win_sz, win_sz_mask);
     } else {
-        peer->tids_next_rel_idx[tid] = idx_end;
+        peer->tids_next_rel_idx[tid] = (u_int16_t)idx_end;
     }
 
     idx_start &= win_sz_mask;
